+++ Assets/Scripts/Game/Player/Models/WaterModel.cs
@@ -6,23 +6,23 @@ namespace Game.Player.Models
 {
     public class WaterModel
     {
-        public ReactiveProperty<int> Value { get; }
-        public ReactiveProperty<int> MaxValue { get; }
+        public ReactiveProperty<float> Value { get; }
+        public ReactiveProperty<float> MaxValue { get; }
 
         public WaterModel(
             IPlayerParametersProvider playerParametersProvider
         )
         {
-            Value = new ReactiveProperty<int>(playerParametersProvider.PlayerParameters.Water);
+            Value = new ReactiveProperty<float>(playerParametersProvider.PlayerParameters.Water);
             MaxValue = new(playerParametersProvider.PlayerParameters.Water);
         }
 
-        public void IncreaseValue(int value)
+        public void IncreaseValue(float value)
         {
             Value.Value = Mathf.Clamp(Value.Value + value, 0, MaxValue.Value);
         }
 
-        public void DecreaseValue(int value)
+        public void DecreaseValue(float value)
         {
             Value.Value = Mathf.Clamp(Value.Value - value, 0, MaxValue.Value);
         }
