+++ Assets/Scripts/Game/Services/Needs/INeedService.cs
@@ -2,7 +2,7 @@ namespace Game.Services.Needs
 {
     public interface INeedService
     {
-        void DecreaseNeed(int value, ENeedType type);
+        void DecreaseNeed(float value, ENeedType type);
         void IncreaseNeed(int value, ENeedType type);
     }
 }
